+++ src/Handler/EasyHandle.php
@@ -109,6 +109,11 @@ final class EasyHandle
     public ?TimeoutException $bodyReadTimeoutException = null;
 
     /**
+     * @var TimeoutException|null Exception during request body rewind timeout.
+     */
+    public ?TimeoutException $bodyRewindTimeoutException = null;
+
+    /**
      * @var \Throwable|null Exception during request body read.
      */
     public ?\Throwable $bodyReadException = null;